+++ tests/processes/test_convert.py
@@ -1,6 +1,7 @@
 """
 Unit tests of functions within :mod:`weaver.processes.convert`.
 """
+from collections import OrderedDict
 from copy import deepcopy
 
 import pytest
@@ -22,7 +23,8 @@ from weaver.processes.convert import (
     is_cwl_file_type,
     json2wps_allowed_values,
     json2wps_datatype,
-    merge_io_formats
+    merge_io_formats,
+    normalize_ordered_io
 )
 from weaver.utils import null
 
@@ -538,3 +540,105 @@ def test_merge_io_formats_wps_overlaps_cwl():
         Format(CONTENT_TYPE_APP_XML),
         Format(CONTENT_TYPE_TEXT_PLAIN),
     ])
+
+
+def test_normalize_ordered_io_with_builtin_dict_and_hints():
+    """
+    Validate that I/O are all still there in the results with their respective contents.
+
+    Literal types should be modified to a dictionary with ``type`` key.
+    All dictionary contents should then remain as is, except with added ``id``.
+
+    .. note::
+        Ordering is not mandatory, so we don't validate this.
+        Also actually hard to test since employed python version running the test changes the behaviour.
+    """
+    test_inputs = {
+        "id-literal-type": "float",
+        "id-dict-details": {
+            "type": "string"
+        },
+        "id-array-type": {
+            "type": {
+                "type": "array",
+                "items": "float"
+            }
+        },
+        "id-literal-array": "string[]"
+    }
+    test_wps_hints = [
+        {"id": "id-literal-type"},
+        {"id": "id-array-type"},
+        {"id": "id-dict-with-more-stuff"},
+        {"id": "id-dict-details"},
+    ]
+    expected_result = [
+        {"id": "id-literal-type", "type": "float"},
+        {"id": "id-dict-details", "type": "string"},
+        {"id": "id-array-type", "type": {"type": "array", "items": "float"}},
+        {"id": "id-literal-array", "type": "string[]"}
+    ]
+    result = normalize_ordered_io(test_inputs, test_wps_hints)
+    assert isinstance(result, list) and len(result) == len(expected_result)
+    # *maybe* not same order, so validate values accordingly
+    for expect in expected_result:
+        validated = False
+        for res in result:
+            if res["id"] == expect["id"]:
+                assert res == expect
+                validated = True
+        if not validated:
+            raise AssertionError("expected '{}' was not validated against any result value".format(expect["id"]))
+
+
+def test_normalize_ordered_io_with_ordered_dict():
+    test_inputs = OrderedDict([
+        ("id-literal-type", "float"),
+        ("id-dict-details", {"type": "string"}),
+        ("id-array-type", {
+            "type": {
+                "type": "array",
+                "items": "float"
+            }
+        }),
+        ("id-literal-array", "string[]"),
+    ])
+    expected_result = [
+        {"id": "id-literal-type", "type": "float"},
+        {"id": "id-dict-details", "type": "string"},
+        {"id": "id-array-type", "type": {"type": "array", "items": "float"}},
+        {"id": "id-literal-array", "type": "string[]"}
+    ]
+    result = normalize_ordered_io(test_inputs)
+    assert isinstance(result, list) and len(result) == len(expected_result)
+    assert result == expected_result
+
+
+def test_normalize_ordered_io_with_list():
+    """
+    Everything should remain the same as list variant is only allowed to have I/O objects.
+
+    (i.e.: not allowed to have both objects and literal string-type simultaneously as for dictionary variant).
+    """
+    expected_result = [
+        {"id": "id-literal-type", "type": "float"},
+        {"id": "id-dict-details", "type": "string"},
+        {"id": "id-array-type", "type": {"type": "array", "items": "float"}},
+        {"id": "id-literal-array", "type": "string[]"}
+    ]
+    result = normalize_ordered_io(deepcopy(expected_result))
+    assert isinstance(result, list) and len(result) == len(expected_result)
+    assert result == expected_result
+
+
+def test_normalize_ordered_io_when_direct_type_string():
+    inputs_as_strings = {
+        "input-1": "File[]",
+        "input-2": "float"
+    }
+    result = normalize_ordered_io(inputs_as_strings)
+    assert isinstance(result, list)
+    assert len(result) == len(inputs_as_strings)
+    assert all([isinstance(res_i, dict) for res_i in result])
+    assert all([i in [res_i["id"] for res_i in result] for i in inputs_as_strings])
+    assert all(["type" in res_i and res_i["type"] == inputs_as_strings[res_i["id"]] for res_i in result])
